pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aristotle-nn"
version = "0.1.0"
description = "Нейросеть на основе текстов Аристотеля"
readme = "README.md"
authors = [
    {name = "AI Developer", email = "ai@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=1.10.0",
    "transformers>=4.18.0",
    "datasets>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
training = [
    "optuna>=3.0.0",
    "ray[tune]>=2.0.0",
    "wandb>=0.13.0",
]
inference = [
    "fastapi>=0.85.0",
    "uvicorn>=0.18.0",
    "prometheus-client>=0.14.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=22.3.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/aristotle-nn"
"Bug Tracker" = "https://github.com/yourusername/aristotle-nn/issues"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py38", "py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "--cov=src"